all functionality of the app working

+++ src/component/Review/ReviewCard.js
@@ -9,21 +9,39 @@ export const ReviewCard = ({ review, handleDeleteReview }) => {
     const getReadableDate = (date) => {
         return new Date(date).toLocaleDateString();
     }
-    return (
-        <div className="review_card">
-            <p className="review_schedule">Name of Hospital: {review.hospital.name} </p>
-            <p className="review_schedule">How easy was to schedule appointment: {review.howEasyWasToScheduleAppointment} </p>
-            <p className="review_schedule">How long did you wait: {review.howLongDidYouWait} </p>
-            <p className="review_schedule">How satisfied are you: {review.howSatisfiedAreYou} </p>
-            <p className="review_schedule"> How would you rate the over all care 1 to 10:{review.howWouldYouRateTheOverAllCare1to10} </p>
-            <div className="review_date">Date {getReadableDate(review.date)}</div>
 
-            <button className="button-edit" type="button"
-                onClick={() => history.push(`/reviews/${review.id}/edit`)}>
-                Edit
-            </button>
-            <button className="button-delete" type="button" onClick={() => handleDeleteReview(review.id)}>Delete</button>
+    const loggedInUser = parseInt(sessionStorage.getItem("medical_appointment_user"));
 
-        </div>
-    )
+
+    if (review.userId === loggedInUser) {
+        return (
+            <div className="review_card">
+                <p className="review_schedule">Name of Hospital: {review.hospital.name} </p>
+                <p className="review_schedule">How easy was to schedule appointment: {review.howEasyWasToScheduleAppointment} </p>
+                <p className="review_schedule">How long did you wait: {review.howLongDidYouWait} </p>
+                <p className="review_schedule">How satisfied are you: {review.howSatisfiedAreYou} </p>
+                <p className="review_schedule"> How would you rate the over all care 1 to 10:{review.howWouldYouRateTheOverAllCare1to10} </p>
+                <div className="review_date">Date {getReadableDate(review.date)}</div>
+
+                <button className="button-edit" type="button"
+                    onClick={() => history.push(`/reviews/${review.id}/edit`)}>
+                    Edit
+                </button>
+                <button className="button-delete" type="button" onClick={() => handleDeleteReview(review.id)}>Delete</button>
+
+            </div>
+        )
+    } else {
+        return (
+            <div className="review_card">
+                <p className="review_schedule">Name of Hospital: {review.hospital.name} </p>
+                <p className="review_schedule">How easy was to schedule appointment: {review.howEasyWasToScheduleAppointment} </p>
+                <p className="review_schedule">How long did you wait: {review.howLongDidYouWait} </p>
+                <p className="review_schedule">How satisfied are you: {review.howSatisfiedAreYou} </p>
+                <p className="review_schedule"> How would you rate the over all care 1 to 10:{review.howWouldYouRateTheOverAllCare1to10} </p>
+                <div className="review_date">Date {getReadableDate(review.date)}</div>
+            </div>
+        )
+
+    }
 }
